requirements.txt
arrow==0.15.7
asgiref==3.2.7
astroid==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
dateparser==0.7.4
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.4.0
django-filter==2.3.0
django-heroku==0.3.1
-e git+https://github.com/telebotter/django-telegrambot.git@5f07ec585086eca9f26d10305eca58fe21c67a6d#egg=django_telegrambot
djangorestframework==3.11.0
future==0.18.2
gunicorn==20.0.4
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.7.5
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.7
pytz==2020.1
regex==2020.5.14
six==1.14.0
sqlparse==0.3.1
timeloop==1.0.2
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
tzlocal==2.1
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
